Finished Endpoints. Improve routing

+++ users/api.py
@@ -11,10 +11,17 @@ from django.shortcuts import get_object_or_404
 from rest_framework import status
 from rest_framework.pagination import PageNumberPagination
 
+from users.permissions import UserPermission
+from rest_framework.viewsets import ViewSet
 
-class UserListAPI(APIView):
+class UserViewSet(ViewSet):
 
-    def get(self, request):
+    serializer_class = UserSerializer
+    pagination_class = PageNumberPagination
+    permission_classes = (UserPermission,)
+
+    def list(self, request):
+        self.check_permissions(request)
         # Instancio paginador
         paginator = PageNumberPagination()
 
@@ -27,7 +34,8 @@ class UserListAPI(APIView):
         # Devolvemos respuesta paginada
         return paginator.get_paginated_response(serialized_users)
 
-    def post(self, request):
+    def create(self, request):
+        self.check_permissions(request)
         serializer = UserSerializer(data=request.data)
         if serializer.is_valid():
             new_user = serializer.save()
@@ -36,26 +44,27 @@ class UserListAPI(APIView):
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
-
-class UserDetailAPI(APIView):
-
-    def get(self, request, pk):
-
+    def retrieve(self, request, pk):
+        self.check_permissions(request)
         user = get_object_or_404(User, pk=pk)
+        self.check_object_permissions(request, user)
         serializer = UserSerializer(user)
         return Response(serializer.data)
 
-    def put(self, request, pk):
+    def update(self, request, pk):
+        self.check_permissions(request)
         user = get_object_or_404(User, pk=pk)
+        self.check_object_permissions(request, user)
         serialiazer = UserSerializer(instance=user, data=request.data)
         if serialiazer.is_valid():
             serialiazer.save()
             return Response(serialiazer.data, status=status.HTTP_200_OK)
         else:
-            return Response(serialiazer.errors,status=status.HTTP_400_BAD_REQUEST)
+            return Response(serialiazer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    def delete(self, request, pk):
-
+    def destroy(self, request, pk):
+        self.check_permissions(request)
         user = get_object_or_404(User, pk=pk)
+        self.check_object_permissions(request, user)
         user.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
